Add map popup styling and refactor pin creation logic
- Introduced a new CSS file for styling map popups, enhancing the visual presentation of pin details.
- Updated the Map component to apply the new popup styling.
- Refactored CreatePin component to utilize Zustand for state management, replacing previous context usage.
- Removed outdated reducer tests to streamline the codebase and improve maintainability.
