+++ client/src/mapPopup.css
@@ -1,0 +1,7 @@
+/* Pin detail popovers on the map */
+.pin-popup .mapboxgl-popup-content {
+  box-shadow:
+    0 12px 28px rgba(0, 0, 0, 0.22),
+    0 4px 12px rgba(0, 0, 0, 0.14);
+  border-radius: 8px;
+}
